move geometry build from Scene to Tile
